Rename prefetcher to start with ROCDL, update description
