+++ compiler/src/iree/compiler/Codegen/LLVMGPU/BUILD.bazel
@@ -102,7 +102,6 @@ iree_compiler_cc_library(
         "LLVMGPULinkExecutables.cpp",
         "LLVMGPULowerExecutableTarget.cpp",
         "LLVMGPUPackSharedMemoryAlloc.cpp",
-        "LLVMGPUPrefetching.cpp",
         "LLVMGPUSelectLoweringStrategy.cpp",
         "LLVMGPUTensorCoreVectorization.cpp",
         "LLVMGPUTensorPad.cpp",
@@ -116,6 +115,7 @@ iree_compiler_cc_library(
         "ROCDLConfigureBufferInstructions.cpp",
         "ROCDLKernelConfig.cpp",
         "ROCDLLowerExecutableTarget.cpp",
+        "ROCDLPrefetching.cpp",
         "ROCDLSelectLoweringStrategy.cpp",
         "TestLLVMGPUQueryMMAPass.cpp",
         "Verifiers.cpp",
