+++ compiler/src/iree/compiler/Codegen/LLVMGPU/CMakeLists.txt
@@ -82,7 +82,6 @@ iree_cc_library(
     "LLVMGPULinkExecutables.cpp"
     "LLVMGPULowerExecutableTarget.cpp"
     "LLVMGPUPackSharedMemoryAlloc.cpp"
-    "LLVMGPUPrefetching.cpp"
     "LLVMGPUSelectLoweringStrategy.cpp"
     "LLVMGPUTensorCoreVectorization.cpp"
     "LLVMGPUTensorPad.cpp"
@@ -96,6 +95,7 @@ iree_cc_library(
     "ROCDLConfigureBufferInstructions.cpp"
     "ROCDLKernelConfig.cpp"
     "ROCDLLowerExecutableTarget.cpp"
+    "ROCDLPrefetching.cpp"
     "ROCDLSelectLoweringStrategy.cpp"
     "TestLLVMGPUQueryMMAPass.cpp"
     "Verifiers.cpp"
